% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.R
\name{allele_frequency}
\alias{allele_frequency}
\title{calculate allele frequency based on genotype frequency}
\usage{
allele_frequency(geno_freq)
}
\arguments{
\item{geno_freq}{list of genotype frequency}
}
\value{
list of allele frequency
}
\description{
calculate allele frequency based on genotype frequency
}
\examples{
allele_frequency(list(AA = 0.01, Aa = 0.18, aa = 0.81))
}
