% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{analyz_frequency}
\alias{analyz_frequency}
\title{Calculate the according frequency of phenotypes and allele}
\usage{
analyz_frequency(counts, pheno_geno_map)
}
\arguments{
\item{counts}{total counts of occurance of each phenotype}

\item{pheno_geno_map}{a map (dataframe) whose keys are phenotype
and values are list of its genes}
}
\value{
list of map, geno frequency, and allele frequency
}
\description{
Calculate the according frequency of phenotypes and allele
}
