% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{clean_lab4_excel}
\alias{clean_lab4_excel}
\title{clean the lab data sheet}
\usage{
clean_lab4_excel(raw)
}
\arguments{
\item{raw}{a raw dataframe from reading the excel}
}
\value{
a cleaned version for computation
}
\description{
this function is form specific
may not appliable if the form changes
}
