% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_growth.R
\name{complete_life_table}
\alias{complete_life_table}
\title{Life Table}
\usage{
complete_life_table(df)
}
\arguments{
\item{df}{an initial life table with
Age in years (x)
Number of Individuals (Nx)
Number of offspring (N x off)}
}
\value{
a Life Table object with everything needed
}
\description{
complete a life table
see PPT 8_PopulationDemographic P5-11
}
\examples{
df <- data.frame(
    x = 0:12,
    Nx = c(90, 39, 33, 30, 29, 27, 16, 8, 5, 5, 3, 2, 0),
    Nxoff = c(0, 2, 22, 45, 19, 149, 11, 0, 0, 11, 0, 0, NA)
)
complete_life_table(df)
}
