% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.R
\name{expect_genotype_frequency}
\alias{expect_genotype_frequency}
\title{expection of one genotype frequence under HWE}
\usage{
expect_genotype_frequency(allele_freq)
}
\arguments{
\item{allele_freq}{list frequence of each allele}

\item{geno}{str genotype}
}
\value{
double exception of genotype frequence
}
\description{
expection of one genotype frequence under HWE
}
\examples{
afreq <- allele_frequency(geno_freq)
expect_genotype_frequency(afreq)('Aa')
}
