% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succession.R
\name{from_lab6_excel}
\alias{from_lab6_excel}
\title{load data from lab6 excel
NOTE not a genaric function}
\usage{
from_lab6_excel(xlsx)
}
\arguments{
\item{xlsx}{string path to excel file}
}
\value{
list of dataframes
}
\description{
load data from lab6 excel
NOTE not a genaric function
}
