% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_growth.R
\name{growth_rate.logit}
\alias{growth_rate.logit}
\title{Logistic Growth Rate}
\usage{
growth_rate.logit(Nt, N0, t)
}
\arguments{
\item{Nt}{population size after time t}

\item{N0}{the initial population size}

\item{t}{units of time}
}
\value{
growth rate
}
\description{
calculate the growth rate under logit growth
}
