% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_growth.R
\name{growth_time.logit}
\alias{growth_time.logit}
\title{Logistic Growth Time}
\usage{
growth_time.logit(Nt, N0, r)
}
\arguments{
\item{Nt}{population size after time t}

\item{N0}{the initial population size}

\item{r}{the intrinsic growth rate}
}
\value{
time takes the population size to grow
}
\description{
calculate the growth time under logit growth
}
