% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.R
\name{is_HWE}
\alias{is_HWE}
\title{check for Hardy-Weinberg Equilibrium for a list of genotypes}
\usage{
is_HWE(genotype_freq, epsilon = 1e-05)
}
\arguments{
\item{genotype_freq}{vector in form of AA, Aa, aa}

\item{epsilon}{threshold for difference}
}
\value{
bool whether it is at HW equilibrium
}
\description{
check for Hardy-Weinberg Equilibrium for a list of genotypes
}
\examples{
is_HWE(list(AA = 0.01, Aa = 0.18, aa =  0.81))
is_HWE(list(AA = 0.25, Aa = 0.5,  aa = 0.25))
is_HWE(list(AA = 0.36, Aa = 0.6,  aa = 0.04))
is_HWE(list(AA = 0.09, Aa = 0.42, aa =  0.49))
is_HWE(list(AA = 0.64, Aa = 0.2,  aa = 0.16))
}
