% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succession.R
\name{panel_coverage}
\alias{panel_coverage}
\title{complete the coverage table for one panel}
\usage{
panel_coverage(a, b)
}
\arguments{
\item{a}{total \% converage}

\item{b}{\% converage on panel surface}
}
\value{
complete table with side information
}
\description{
complete the coverage table for one panel
}
\examples{
A_3_months <- c(17, 1 / 2, 10 + 1 / 4, 1 / 4, 1 / 2) / 25
B_3_months <- c(17, 1 / 2, 5 + 1 / 4, 1 / 4, 0) / 25
panel_coverage(A3, B3)

}
