% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_growth.R
\name{pop_growth.logit}
\alias{pop_growth.logit}
\title{Logistic Growth Population Size}
\usage{
pop_growth.logit(N0, r, t, K)
}
\arguments{
\item{N0}{the initial population size}

\item{r}{the intrinsic growth rate}

\item{t}{units of time}
}
\value{
predicted population size
}
\description{
calculate the  predicted population size under logit growth
}
