% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_growth.R
\name{pop_prowth.exp}
\alias{pop_prowth.exp}
\title{Expoential Growth Population Size}
\usage{
pop_prowth.exp(N0, r, t)
}
\arguments{
\item{N0}{the initial population size}

\item{r}{the intrinsic growth rate}

\item{t}{units of time}
}
\value{
predicted population size
}
\description{
calculate the  predicted population size under exponential growth
}
