% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{predict.evolution}
\alias{predict.evolution}
\title{Make prediction aboiut evolution on the current generation}
\usage{
\method{predict}{evolution}(generation)
}
\arguments{
\item{generation}{evolution object}
}
\value{
expected geno frequency table of next generation
}
\description{
Make prediction aboiut evolution on the current generation
}
