% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction_curve.R
\name{rarefaction_plot}
\alias{rarefaction_plot}
\title{Rarefaction Plot}
\usage{
rarefaction_plot(df, title = "")
}
\arguments{
\item{df}{A two column dataframe that contains
\# of new species and \# of individuals in order}

\item{title}{A name of the plot}
}
\value{
a ggplot graph object
}
\description{
make a rarefaction plot based on a two column data frame
}
\examples{
rf <- read.csv("../data/lab1_rarefaction.csv", header = F)
makeRarefactionPlot(rf[, 1:2], title = "Group 1")
}
