% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{summary.lab4}
\alias{summary.lab4}
\title{add mean, sd, se to lab data}
\usage{
\method{summary}{lab4}(data)
}
\arguments{
\item{data}{a cleaned dataframe (from clean_lab4_excel)}
}
\value{
mean, standard deviation, standard error
}
\description{
add mean, sd, se to lab data
}
\examples{
summary(clean_lab4_excel(raw_df))
}
