% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succession.R
\name{summary.succession}
\alias{summary.succession}
\title{summary of succession (lab6 data)}
\usage{
\method{summary}{succession}(df, age)
}
\arguments{
\item{df}{dataframe}

\item{month}{age of panel}
}
\value{
summary dataframe
}
\description{
summary of succession (lab6 data)
}
