% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{test_treatment}
\alias{test_treatment}
\title{Test for a Treatment}
\usage{
test_treatment(data, treatement, components, alpha = 0.05)
}
\arguments{
\item{data}{a cleaned dataframe (from clean_lab4_excel)}

\item{treatement}{string the name of this treatement}

\item{components}{list of string components names to compare}

\item{alpha}{float Significance Level}
}
\value{
a complete table of test results
}
\description{
t test for a treatment by given compoents
}
