% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{true_diversity_index}
\alias{true_diversity_index}
\title{D of an environment (multiple samples)}
\usage{
true_diversity_index(df)
}
\arguments{
\item{df}{each row is the number of each species in the sample}
}
\value{
true diversity index of environment
}
\description{
calculate true diversity index of multiple sample
see lab 1 for example
}
\examples{
df <- read.csv("../data/lab1.csv")
true_diversity_index(df)
}
