% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{true_diversity_index.sample}
\alias{true_diversity_index.sample}
\title{D of one sample}
\usage{
true_diversity_index.sample(ns)
}
\arguments{
\item{ns}{number of each species in the sample}
}
\value{
true diversity index of this sample
}
\description{
calculate true diversity index of one sample
see study guide 1 #12
PPT 3_Biodiversity P7
}
\examples{
true_diversity_index.sample(c(4, 2, 3, 11))
}
