% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{plot_mean_err}
\alias{plot_mean_err}
\title{make plot on the component measurements}
\usage{
plot_mean_err(component_df, title = "")
}
\arguments{
\item{component_df}{dataframe of one component}

\item{title}{string name of the plot}
}
\value{
ggplot graph object
}
\description{
make plot on the component measurements
}
\examples{
Mean	    SD	    SE	    treatment	morph
22.44	8.04	2.32	No	        0
14.55	2.68	0.77	No	        1
22.06	12.29	3.55	Intramorph	0
17.03	4.53	1.31	Intramorph	1
16.72	3.97	1.15	Intermorph	0
17.67	7.27	2.10	Intermorph	1
}
