% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{test_treatment}
\alias{test_treatment}
\title{Test for a Treatment}
\usage{
test_treatment(memo, treatement, components, df, alpha = 0.05)
}
\arguments{
\item{memo}{a cleaned dataframe (from summerize_lab4)}

\item{treatement}{string the name of this treatement}

\item{components}{list of string components names to compare}

\item{alpha}{float Significance Level}
}
\value{
a complete table of test results
}
\description{
t test for a treatment by given compoents
}
\examples{
summ <- summary(clean_lab4_excel(raw_df))
test_treatment(summ$data, treatments[i], components, summ$df)
}
